Add code block support to articles (#2)
## Background

I want to be able to show code examples in my articles. To achieve a
nice developer experience it should look and feel like a coding example
in the Sanity Studio as well.

## Solution

Added the `@sanity/code-input` package which basically does most of what
I want out of the box! Created a type for it, and added a new field to
the content array in the `Article` type to have a list of them. This is
to future proof, as there probably will be a point where I want
side-by-side code blocks for comparison!