Add rich image type (#4)
## Background

We need to be able to show some pictures. It would be very boring if
text and code are the only forms of content. Also, the image should
support some metadata, like captions and alternative texts for
accessibility.

## Solution

Created a rich image type. This is a Sanity image with three added
fields that are pretty standard for any picture:
* alt text
* caption
* credits

Also enabled hotspot as it is a nice QOL feature from Sanity. To use
this in articles I made a list type that can have many images.