Add proper validation for typegen types (#5)
## Background

To get correct types for the frontend, we can use Sanity's TypeGen
feature, where you extract the schema of your Sanity Studio, and then
generate types from this schema. This though, will mostly end up
nullable if you don't specify that you want enforcing rules, and have
used proper validation for those types!

## Solution

Tested a bit, and found the fields that missed validation to be properly
parsed in the frontend.